[이동혁] 문제은행3 숙제(12.30)

+++ DongHyeokLee/java/6th/src/Ans7.java
@@ -30,7 +30,7 @@ public class Ans7 {
         float percent;
 
         //int[] people = new int[END];
-        int[] emp = new int[EMP_NUM];
+        float[] emp = new float[EMP_NUM];
 
 
 
@@ -38,7 +38,7 @@ public class Ans7 {
             //people[num] = pay;
             for(int j = 0; j < EMP_NUM; j++) {
                 //int rate = (int) (Math.random() * range + MIN); // 0 <= rate < 1 0<= rate < 9 1<= rate < 10 1/100
-                percent = (float) ((int) (Math.random() * range + MIN / BIAS) * PERCENT);
+                percent = (float) ((int) (Math.random() * range + MIN) / BIAS) * PERCENT;
                /* rise = pay * rate / 100; // 계속 0으로 작동
                 risePay = (int) pay + rise; // 인상된 연봉에서 또 인상?? 이렇게 되면 소수점 까지 뜨게됨 근데 오른거에서 오르는게 당연하자나 그럼 doulbe형을 int형으로 바꿔야되나
                 pay = risePay; // pay를 final로 고정시키니 작동이 안됐음*/
